refactor: Calculator, Expression 테스트 분리
